Split util funs from vproj/projects.go to vproj/util.go and refactored tarFromVorteilProject function
